mmc: fsl_esdhc: add workaround for non-removable card of esdhc-2
The esdhc-2 usually uses some on-board memory devices such as eMMC
card or SDIO wifi module, and it doesn't support SDHC_CD_B. So we
could only assume it always has a card instead of detecting SDHC_CD_B
status. This patch is to add workaround for these non-removable
cards which are used by esdhc-2.

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>