sf: DSPI: limit the flash size for N25Q512A to 16MB
There is an issue for DSPI flash on LS2088ARDB.
The DSPI flash on LS2088ARDB is N25Q512A, it’s a 64MB flash.
So far, the SPI framework in uboot doesn’t support 4Byte address mode.
So uboot can’t support N25Q512A with 16M+ flash.
That means now the DSPI on LS2088ARDB with N25Q512A can’t work.

We will deliver the support in two phases:
•	Basic DSPI support (upto 16M)
•	Advanced DSPI support (64M, 4byte address mode)

So here is a temp workaround to limit the flash size for N25Q512A to
16MB.

Signed-off-by: Yuan Yao <yao.yuan@nxp.com>