pci: layerscape: Adjust PCIe bus number assignment
Only the first PCIe controller is assigned to bus number 0.
The second/third PCIe controllers may be assigned to bus
number 2 / 3. When a PCIe device is inserted to these slots,
the bus number is assigend followed the controller bus number,
but the bus number set to the register enquals
device bus - controller bus. So fix it.

Signed-off-by: Minghuan Lian <Minghuan.Lian@nxp.com>
Signed-off-by: Liu Gang <Gang.Liu@nxp.com>