ls1088a: add VID support for QDS and RDB platforms
This patch adds the support for VID on LS1088AQDS and LS1088ARDB systems.
It reads the fusesr register and changes the VDD accordingly by adjusting the
voltage via LTC3882 regulator.
This patch also takes care of the special case of 0.9V VDD. In that case,
it also changes the SERDES voltage by disabling the SERDES, changing the SVDD
and then re-enabling SERDES.

Signed-off-by: Raghav Dogra <raghav.dogra@nxp.com>